TelegramBot/requirements.txt
python-dotenv==1.0.0
aiogram==2.25.1


gino==1.0.1
SQLAlchemy==1.3.24
asyncpg==0.27.0